Lock down the django-multiselectfield dependency for older releases.
A new django-multiselectfield recently came out, and despite it being a
micro version bump and listing Python 2.7 compatibility, it makes use of
Python 3 type hints. This breaks older builds.

Supported releases of Review Board are not affected by this, but in an
effort to keep CI working and to avoid surprises if we ever need to
supply an older build to a customer, we're now capping this dependency
to an explicit 0.1.12 release.

Testing Done:
Tested a new environment and verified I got the 0.1.12 release.

Reviewed at https://reviews.reviewboard.org/r/14092/