Fix text e-mails for review replies.
The template for text e-mails for review replies was inadvertantly
including the `.html` version of the content template instead of `.txt`.
This was causing the message to send HTML code, but marked as plain
text.

Testing Done:
Loaded the preview-email page and saw that before the fix it showed HTML
code and after it showed the expected text content.

Reviewed at https://reviews.reviewboard.org/r/14111/