Add release notes for Review Board 7.0.2.
The major change in this release is the new review experience for
administrators, helping them participate in the review process like any
other user but jump into an administrator role when needed.

There are also some important compatibility fixes for Mercurial and
Perforce, and a handful of other bug fixes and behavioral improvements.

Screenshots for the new administrator experience will be added in a
later update to this change.

Testing Done:
Built the docs. Checked for build errors, obvious bad spelling issues,
and bad links.

Reviewed at https://reviews.reviewboard.org/r/14108/