Fix regression where change description file attachment thumbnails disappear.
Release 7.0.2 introduced a regression where file attachment thumbnails
in change descriptions dissapeared. The cause of the regression existed
before, but didn't cause any actual problems until it was exposed by the
new "admin unpublished draft view" changes.

In those new changes we grab data from the `reviewRequestEditor` option on
a `fileAttachmentThumbnailView` to query some state. The `reviewRequestEditor`
option is a mandatory option, so the `fileAttachmentThumbnailView` expects it
to be defined. When we create file attachment thumbnail views in our change
descriptions however, we weren't defining it.

This change makes sure we're defining the expected options on file
attachment thumbnail views everywhere in the code.

Testing Done:
- Saw that file attachment thumbnails appear correctly in change descriptions,
  review requests, and review UIs.

Reviewed at https://reviews.reviewboard.org/r/14113/