Fix the ability to run Storybook on Review Board 7.
The support for Storybook for UI component testing broke since it was
introduced. Part of this was missing dependencies that are needed for
the current configuration. Part of it was a JavaScript file that's no
longer provided. Part of it was references to Ink that weren't provided
by package exports, which is being fixed in Ink.

With this change and the Ink export fix, Storybook should now be usable
again for UI testing.