Use a <table> for the related users option list.
For the set of selected users when using a related users selector
widget, things were wrapping in terrible ways if there were long
usernames or real names. This change builds on a change in Djblets to
use a <table> for the layout of the options instead of a <ul>.

While in here, I also decided to make the widget a bit bigger, because
it was pretty small even for reasonable-sized desktop windows.

Testing Done:
Created some users with very long names and added them to a group. Saw
that the users selector widget no longer looked terrible.

Reviewed at https://reviews.reviewboard.org/r/14135/