Add a mobile mode attribute to the review request editor view.
A `inMobileMode` attribute already exists on the base `PageView` class for
letting views know whether they're being viewed on a mobile device. This
change plumbs that attribute through to the `ReviewRequestEditorView`. This
will be useful in an upcoming change for fixing file attachment actions
on mobile, and is generally useful if we need to do any mobile specific
behaviour on the review request editor view or any other views attached
to it in the future.

Testing Done:
- Ran JS unit tests.
- Used in an upcoming change.

Reviewed at https://reviews.reviewboard.org/r/14126/