Fix the display of diff comments of a review on mobile.
For diff comments that were on new or deleted files, the heading for the file
would spill outside of the review box. This is fixed by adding a
`display: grid` to the diff table and to the filename row in the table for
diffs that appear in review comments. This problem didn't appear for diff
comments on existing files because we were already specifically giving
them a grid display.

Testing Done:
- Looked at diff comments on new files in mobile and desktop.
- Looked at diff comments on existing files in mobile and desktop.
- Looked at the diff viewer in mobile and desktop, made sure it was the
  same as before.

Reviewed at https://reviews.reviewboard.org/r/14159/