Move the Diff Complexity Icon story into a "Diffs" category.
In preparation for new component stories for Storybook, this change does
a bit of preemptive reorganization, moving the Diff Complexity Icon into
a "Diffs" sub-category of "Components". This will help separate it from
other stories built in the future.

Testing Done:
Re-ran Storybook and saw that the component was in the new category.

Reviewed at https://reviews.reviewboard.org/r/14143/