Improve the look of the unified banner's buttons on mobile.
Depending on the screen size, the publish/discard buttons on the unified
banner could be squished up against the left side of the screen instead
of being at the top right corner of the banner. This change adds a
left margin to the buttons on mobile to make things look better.

Testing Done:
Made sure the buttons looked good in the different draft modes and
tested tapping them.

Reviewed at https://reviews.reviewboard.org/r/14131/