Add Dark Mode for the Admin UI.
Review Board 7.1 will ship with Dark Mode enabled all throughout the
product. We introduced it as a beta, opt-in feature in 7.0, but we
didn't have it universally, as we didn't have time to convert the Admin
UI or some of the components in Djblets. Since then, we've made progress
on pretty much every component the Admin UI depends on.

We now support Dark Mode all throughout. This includes all the Admin
settings pages, dashboard widgets, security checks, and more. Colors
were converted to Ink color palettes, buttons to Ink buttons, and
various styles fixed.

An upcoming change will enable Dark Mode for all users based on system
settings.

As a note, the support banner in the Admin Dashboard is bright, but there
are plans to make further changes specifically to this area in an
upcoming change, so it will remain as alert styling for now.

Testing Done:
Tested all pages in the Admin UI in dark mode and in light mode, checking
for any visual issues.

Screenshots on the review request are included just for notable pages.
Most use standard dark mode styling and color palettes and aren't notable.

Reviewed at https://reviews.reviewboard.org/r/14166/