Remove legacy padding from the filediff admin styles.
FileDiffs in the admin site (either on the FileDiff page or inline in
the DiffSet page) were showing with a bunch of extra padding. This was
due to some old styles which were added long before we had our current
form layout which uses table-cell display to align everything. This
change removes that extra padding.

Testing Done:
Viewed filediffs in the admin site and saw that there was no longer
a big extra space next to the unified diff content.

Reviewed at https://reviews.reviewboard.org/r/14186/