Add some styling for the rb-c-form-field__disabled-reason element.
Djblets KeyValueForms have a nifty feature where we can annotate why a
field is disabled. We don't use this very much, so this wasn't really
styled at all.

This change adds some styling for that to show similar to help text, but
with an alert/error icon.

Testing Done:
Viewed a form that used the disabled reasons and saw the new style.

Reviewed at https://reviews.reviewboard.org/r/14192/