Fix migration of legacy SCMTools.
When we migrate a database, we migrate legacy SCMTool entrypoints,
creating a map of their class paths to names. Unfortunately, this code
regressed when we moved to the modern importlib-based entrypoints, as we
were accessing old attribute names.

This is a simple change that switches to the modern attributes, fixing
migration.

Testing Done:
Verified this fixed a migration of an old database that had some
entrypoint-based SCMTools.

Reviewed at https://reviews.reviewboard.org/r/14215/