Fix migration of already-migrated SCMTool IDs in integration conditions.
During upgrade, we attempt to convert any SCMTool primary key IDs in
conditions found in integration configurations. If we're dealing with
any partially-migrated conditions, we could end up crashing, since our
assumption that an entry is a primary key will fail.

We now safeguard against this when collecting IDs for the migration map
and then again when we update the values, allowing non-integer IDs to
remain as-is.

Testing Done:
Unit tests passed.

Verified this fixed an issue I encountered in some partially-migrated state.

Reviewed at https://reviews.reviewboard.org/r/14216/