Use Ink colours for flags and comments.
This change adds new `--rb-p-flag-draft-*` and `--rb-p-flag-published-*`
definitions to our standard colour palette. These should be used for flags
that indicate whether an item is a draft or published. The colours either
exactly match the old ones that we used or are very close. We update the
following items to use these new definitions:

* Comment flags
* Comment tooltips
* Image region comments
* Datagrid draft labels
* File attachment draft labels

This change also fixes a syntax mistake in the definition for
`--rb-theme-rich-text-code-fg`, and pulls out a couple variable
definitions for `--rb-theme-rich-text-code-bg` to make code blocks
in comment tooltips look nicer.

Testing Done:
Looked at the comments, datagrid labels and file attachment labels on
darkmode and lightmode.

Reviewed at https://reviews.reviewboard.org/r/14195/