Fix issues with large images in the diff viewer.
This fixes a couple issues that appear when images in the diff viewer
had a larger width than their container. The header that contained the
zoom controls would match the width of the image, so the zoom controls
would be cut off and inacessible. Secondly, there was no scroll bar, so
you wouldn't be able to see the full image when it overflowed in the
container. These issues only happened for large images in the diff
viewer, images in their individual review UI page were fine.

Testing Done:
- Looked at images in the diff viewer (smaller than the container and
  bigger).
- Looked code diffs in the diff viewer and saw that they looked the same
  as before.
- Looked at images in their review UI and made sure there were no issues.

Reviewed at https://reviews.reviewboard.org/r/14228/