Make the comment dialog look better on mobile.
Currently, the comment dialog looks very squished when viewing it on mobile
with published comments. This change fixes this by stacking the published
comments box on top of the draft comment form, instead of side by side. We
now allow the dialog to be dragged from the "Other comments" header in
addition to the draft form header. It's more natural to try to drag the
top "Other comments" header on mobile when trying to move the dialog.

This also updates the z-index for the prev and next file attachment nav
buttons, so that it has a lower z-index than the comment dialog.

Testing Done:
Tested the following on mobile and on desktop:
- Creating pure draft comments.
- Creating comments that are on the same region as an existing comment on
  an image.
- Creating comments that are on the same region as an existing comment on
  a diff.
- Creating comments that are on the same region as an existing comment on
  a markdown file.
- Clicking on an existing comment while not logged in.
- Moving and resizing the comment dialog.

Reviewed at https://reviews.reviewboard.org/r/14188/