Allow comment thumbnails to scroll on overflow.
If a comment thumbnail is to big for the browser (which happens
often on mobile), the image would spill outside of the review container.
This changes it to allow scrolling on overflow instead.

Testing Done:
Tested viewing comment thumbnails that were too wide for the browser. Saw
that it was contained in the review container and could be scrolled. Saw
that the scrollbar only shows up when needed.

Reviewed at https://reviews.reviewboard.org/r/14251/