Fix the "watched list" resource.
The watched list resource is a special resource that exists only to link
to the other watched item resources. It was completely broken, returning
HTTP 405, due to a bad implementation of our singleton resource pattern.

This change fixes that up to use the correct method name, and
reimplements the body of it to skip the bulk of what the base `get_list`
implementation would have done.

Testing Done:
Fetched the watched list resource and got a response with links instead
of an error.

Reviewed at https://reviews.reviewboard.org/r/14258/