Fix scrolling the list of other comments in the comment dialog.
When we implemented the fix to help keep the comment dialog from
scrolling out of bounds when attempting to scroll part of the comment
dialog's UI, we only enabled scrolling for the text area. Attempting to
scroll the list of other comments on a line no longer worked, as the
event would get swallowed.

We now make an exception for this part of the UI, allowing it to scroll
just like the text area.

Testing Done:
Verified I could scroll both the text area and the list of other comments,
but otherwise scroll events did not bubble up to the page.

Reviewed at https://reviews.reviewboard.org/r/14277/