Fix up some references to a non-existent module.
There's no such thing as `reviewboard.reviews.views.detail`, but we were
referencing it in a couple places. For the code reference, this was
supposed to be `reviewboard.reviews.views.review_request_detail`, and
for the 7.0.2 release notes, this was supposed to be
`reviewboard.reviews.detail`.

Testing Done:
Built docs.

Reviewed at https://reviews.reviewboard.org/r/14354/